# docs/api.yml
openapi: "3.0.3"
info:
  title: FinanceApi
  version: "1.0"
servers:
  - url: http://localhost:8080
security:
  - BearerAuth: []
...
paths:
...
  /auth/sign-in/google:
    post:
      tags: [auth]
      security: []
      summary: >
        Login or register with a Google account.
        Accepts a Google id_token obtained from the Google Sign-In SDK on the client.
        Three cases are handled automatically:
        (1) googleId already known — direct login;
        (2) email already exists but not yet linked — googleId is attached and login proceeds;
        (3) unknown email — new account created automatically (no password set).
        Returns the same JWT format as POST /auth/sign-in, valid for 10 hours.
      operationId: signInWithGoogle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleSignInCredentials"
      responses:
        "200":
          description: >
            Authenticated — returns account info and JWT token.
            The account may have been created automatically if it did not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignInResult"
        "400":
          description: >
            idToken is missing, empty, or rejected by Google
            (expired, wrong audience, tampered signature).

  # ─── HEALTH ───────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT token obtained from POST /auth/sign-in. Valid for 10 hours.
        The token embeds id, username and email.
        All protected routes verify that the accountId in the path matches the token's id.
...
  schemas:
...
    GoogleSignInCredentials:
      type: object
      required: [idToken]
      properties:
        idToken:
          type: string
          minLength: 1
          description: >
            Google id_token obtained from the Google Sign-In SDK on the client
            (Android, iOS, or Web). This token is verified server-side against
            GOOGLE_CLIENT_ID using the google-auth-library.
...
    SignUpResult:
      type: object
      description: Account info returned after sign-up or embedded in SignInResult. Password is never returned.
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
    SignInResult:
      type: object
      properties:
        account:
          $ref: "#/components/schemas/SignUpResult"
        token:
          type: string
          description: >
            JWT token valid for 10 hours.
            Use as Bearer in the Authorization header for all protected routes.
            Embeds id, username and email.

    # ── Health ────────────────────────────────────────────────────────────────
...